update drop file

+++ operations/drop.py
@@ -1,9 +1,10 @@
-import sys
-
-sys.path.append("..")
-from config import DEFAULT_TABLE
-from logs import LOGGER
 import shutil
+from logs import LOGGER
+from config import DEFAULT_TABLE
+import sys
+import os
+sys.path.append("..")
+
 
 def do_drop(table_name, milvus_cli, mysql_cli):
     if not table_name:
@@ -14,7 +15,8 @@ def do_drop(table_name, milvus_cli, mysql_cli):
         status = milvus_cli.delete_collection(table_name)
         images_path = mysql_cli.search_image_path_by_table_name(table_name)
         # print("images_path", images_path)
-        shutil.rmtree("/".join(images_path[0].split("/")[:-2]))
+        if os.path.exist(images_path[0].split("/")[:-2]):
+            shutil.rmtree("/".join(images_path[0].split("/")[:-2]))
         mysql_cli.delete_table(table_name)
         return status
     except Exception as e:
